Decisions

Documentation for Decisions.

Decisions.Decisions

Decisions.bayes_strategy_loss

Decisions.expected_strategy_loss

Decisions.expected_strategy_loss

Decisions.Decisions — Module

Functions for expected utility calculations

Decisions.bayes_strategy_loss — Method

bayes_strategy_loss(SL::StrategyLossMatrix, i::Int, OM::ObservationMatrix; prior=nothing)

Compute the Bayes risk of strategy i under a given prior. If prior is nothing, it assumes a uniform prior over states.

Decisions.expected_strategy_loss — Method

expected_strategy_loss(L::LossMatrix, OM::ObservationMatrix)

Compute the risk (expected loss) for each strategy and each state. Returns a NamedArray where rows are named by the strategies and columns are named by states.

Decisions.expected_strategy_loss — Method

expected_strategy_loss(SL::StrategyLossMatrix, OM::ObservationMatrix)

Compute the risk (expected loss) for each strategy and each state using a tensor-based approach. Returns a NamedArray where rows are named by the strategies (tuples in SL.P) and columns are named by states.