Conditional distributions

Decisions.jl provides an interface for conditional distributions (CDs), ConditionalDist, used to implement the nodes of decision networks. Once can draw samples from a conditional distribution, evaluate its PDF, get its support, and so forth, all conditioned on named random variables, so long as the corresponding functions are implemented (which need not be the case).

DecisionNetworks.ConditionalDist — Type

ConditionalDist

A conditional distribution: a stochastic mapping from a set of conditioning variables K to a random variable of type T.

The conditional distribution interface

Functions defined on conditional distributions

Conditional distributions are subclasses of ConditionalDist{K, T} which implement some or all of the following functions:

Function	Default	Description

support	TypeSpace{T}	Support of a CD, possibly conditioned random variables.

rand!	Base.rand(...)	Sample from a CD in place if possible.

pdf	exp(logpdf(...))	Give probability [density] of an output value.

conditions	K	Tuple of Symbol names of random variables conditioning a CD.

Base.eltype	T	Output type of values generated according to a CD.

fix	FixedDist(...)	Fix value of a conditioning variable of a CD for a new CD.

Base.rand	none	Sample from a CD, conditioned on RVs as keyword arguments.

logpdf	none	Give log-probability [density] of a output value.

To be considered valid, all the above functions defined for ConditionalDists must be pure (with the exception of the in-place functions, which may only modify their destination argument.)

Warning

Decisions.jl can't check if your ConditionalDists are pure functions. If they are not, it's impossible to make accurate guarantees about related Decisions objects, which can result in unexpected or incorrect behavior from solvers.

Random variable values are keyword arguments

In functions that accept them, the values of random variables that condition a ConditionalDist are given as keyword arguments (where the keyword is the name of the random variable). Be aware of these consequences:

Conditioning variables are orderless, matching their mathematical presentation (i.e., $p(x| A=a, B=b) = p(x | B=b, A=a)$).

Functions like support can be defined for arbitrary combinations of conditioning

variables, or none at all. Use rv=missing as the default keyword value in this case and check for presence with ismissing.

You cannot dispatch the above functions on conditioning variables.

DecisionNetworks.support — Function

support(cd::ConditionalDist{K, T}; kwargs...)

Give the support Space of the distribution when the conditioning variables K take on the values in kwargs (if it is present), or a superspace containing it.

If no conditioning variables are provided, give the joint support Space over all conditionings. Optionally, if some, but not all, conditioning variables are provided, the joint support Space over the remaining conditioning variables should be returned.

By default, returns a TypeSpace{T}.

Random.rand! — Function

rand!(rng=default_rng(), cd::ConditionalDist{K, T}, dest::T; kwargs...) where {K, T}

Sample from a conditional distribution (in place using dest if possible), with the conditioning variables taking on the values in kwargs.

Returns dest if in-place modification is successful; otherwise, returns a new instance. By default, never modifies in place and defers to rand.

DecisionNetworks.pdf — Function

pdf(cd::ConditionalDist, x; kwargs...)

Gives the probability or probability density of a random variable distributed according to cd with the value x, given values of conditioning variables in kwargs.

Equivalent to cd(x; kwargs...)

DecisionNetworks.conditions — Function

conditions(cd::ConditionalDist)

Report the names of the conditioning variables of cd as a tuple of Symbols.

conditions(dn::DecisionGraph, s::Symbol)
conditions(dn::DecisionNetwork, s::Symbol)

Give the names of the conditioning variables of s in dn (including any indexing variables).

Base.eltype — Method

Base.eltype(::ConditionalDist{K, T}) where {K, T}

When applied to a ConditionalDist, give T, the type of values produced by the distribution.

Base.rand — Method

rand(rng=default_rng(), cd::ConditionalDist; kwargs...)

Sample from a conditional distribution, given values of conditioning variables in kwargs.

Equivalent to cd(; kwargs...).

DecisionNetworks.logpdf — Function

logpdf(cd::ConditionalDist, x; kwargs...)

Gives the natural logarithm of the probability or probability density of the random variable distributed according to cd having the value x, given values of conditioning variables in kwargs.

DecisionNetworks.fix — Function

fix(cd::ConditionalDist; fixed_args...)

Fix any number of variables conditioning a distribution to particular values, returning a new conditional distribution conditioned on the variables that remain.

By default produces a FixedDist.

Statistical syntax

ConditionalDists are callable, and refer to rand (when called with only keyword arguments) and pdf (when called with exactly one argument as well), i.e.

my_dist = MyConditionalDist()
my_dist(; a, b)   # same as rand(my_dist; a, b)
my_dist(x ; a, b) # same as pdf(my_dist, x; a, b)

This mimics the statistical syntax $x \sim p_{\textrm{my_dist}}(\cdot | a, b)$ and $p_{\textrm{my_dist}}(x | a, b)$ respectively, with the semicolon replacing the vertical bar.

Terminal states (and other outputs)

Sometimes a distribution is provided with conditions that represent an exceptional, simulation-halting "terminal" case (i.e., a "terminal state"). Under these conditions, distributions are expected to return the unique value terminal, which is not included in the distribution support.

The Terminality trait can be used to signal that a distribution in a DN should never produce terminal.

DecisionNetworks.Terminal — Type

struct Terminal end

Type of the unique value representing the output of a decision node as being terminal or otherwise exceptional.

DecisionNetworks.terminal — Constant

terminal

Unique value representing the output of a decision node as being terminal or otherwise exceptional. The singleton instance of type Terminal.

DecisionNetworks.isterminal — Function

isterminal(x)

Return true if and only if x === terminal, and false otherwise, in the style of isnothing.

Defining conditional distributions

There are several ways to define and instantiate ConditionalDists.

Explicit conditional distributions

Of course, one can simply write a subclass of ConditionalDist:

using Random
struct MyCD <: ConditionalDist{(:s, :a), Float64}
    val::Float64
end

function Random.rand(rng, cd::MyCD; s, a)
    s+a+cd.val
end

mycd = MyCD(3)
mycd(; s=1, a=2)

6.0

Todo

The rng argument is required. In the future rng should be moved to a special keyword argument.

This is a generative-only distribution; no PDF is known, which is perfectly valid in Decisions.jl. In general, one can define only the functions of a CD that are known. (Another common case is when the support of a distribution is known, but not a PDF or a generative function - this is all that is known about a decision node in a DN before a decision is actually made. See UndefinedDist.)

@ConditionalDist

Fully defining a decision network requires a lot of conditional distributions, and fully specifying them as their own types is tedious, so a macro @ConditionalDist is available.

DecisionNetworks.@ConditionalDist — Macro

@ConditionalDist

Generate an anonymous conditional distribution definition from a set of functions.

The functions can be any of those in the ConditionalDist interface and must be named accordingly, with the ::ConditionalDist argument omitted. They can be defined with the function keyword or in the compact style. The names of conditioning variables are automatically inferred from the names of keyword arguments.

The rng argument is mandatory in functions that use it.

Technically, @ConditionalDist defines an anonymous conditional distribution which is a unique subtype of ConditionalDist.

DecisionNetworks.AnonymousDist — Type

AnonymousDist

An anonymous distribution: one defined with @ConditionalDist.

It carries its implementation as higher order functions, so every AnonymousDist has a unique type (not unlike anonymous functions).

Anonymous conditional distributions can be defined inside functions and other places where structs cannot usually be defined. They can also include closures over external variables. This substantially condenses CD definitions - for instance, the previous example can be rewritten as:

    MyCD(val) = @ConditionalDist Float64 begin
        rand(rng; s, a) = s+a+val
    end
    mycd = MyCD(3)
    mycd(; s=1, a=2)

6.0

Structural conditional distributions

DecisionNetworks.jl provides a few very basic distributions explicitly which have important structural meaning:

DecisionNetworks.UndefinedDist — Type

UndefinedDist

A distribution that has no PDF or sampling defined. Only its support is known.

DecisionNetworks.CompoundDist — Type

CompoundDist{K, T, Ki} <: ConditionalDist{K, T}

A distribution which implements P(⋅ | idx, ...) using a Tuple of P(⋅ | ...) distributions (where idx maps into that tuple).

Useful for merging behavior of multiple agents into a single distribution.

DecisionNetworks.CollectDist — Type

CollectDist <: ConditionalDist

A deterministic conditional dist which simply stacks its inputs as a Tuple.

DecisionNetworks.FixedDist — Type

FixedDist <: ConditionalDist

A conditional distribution that wraps another, maintaining values of fixed variables; also, the default output of fix.

Forwards most standard ConditionalDist functions to its underlying distribution, conditioned on the inputs and the fixed variables.

DecisionNetworks.UniformDist — Type

UniformDist{K, T} <: ConditionalDist{K, T}

A discrete uniform distribution: selects elements from its finite support with equal probability.

Conditional distributions from Distributions.jl

Todo

Not currently implemented.

Objects that convert to ConditionalDists

Some basic objects can be interpreted as ConditionalDists with convert:

Functions are interpreted as sample generators and form generative-only distributions. The conditioning variable names and the output type must be specified by the type parameters {K, T} on the target ConditionalDist{K, T}.

Spaces are interpreted as UndefinedDists which return the given space their unconditioned support. Only the conditioning variable names {K} on the target ConditionalDist{K} need be defined; the output type T is inferred from the space.