Decision networks

The fundamental unit of problem description in Decisions.jl is the (dynamic) decision network or (D)DN (variously related to or synonymous with "influence diagram," "generalized Bayesian network," and "decision diagram"). A decision network is a directed acyclic graph where each node is the distribution of a random variable conditioned on some other nodes in the network.

In Decisions.jl, DNs consist of two components:

A decision graph, which gives the abstract definition of the relationships between random variables in the network, and

An implementation, which provides some of the conditional distributions in the network. Nodes with no implemented distribution are considered decision nodes.

The decision graph lives in type space: it is given by the Type{<:DecisionNetwork} (aliased to DecisionGraph). This allows Decisions.jl to perform computation about the structure of a DN at compile time. DecisionNetwork instances additionally carry implementation information.

DecisionNetworks.DecisionNetwork — Type

DecisionNetwork

Representation of a decision network based on an underlying directed acyclic graph.

DecisionNetworks.DecisionGraph — Type

const DecisionGraph = Type{<:DecisionNetwork}
DecisionGraph(nodes, dynamic_pairs=nothing, ranges=nothing)

A decision graph: the graph structure of a decision network, with none of the distributions implemented.

If dynamic_pairs or ranges is nothing, the corresponding DecisionNetwork type parameter is left unspecified. (This way, it is possible to define a DecisionGraph without knowing the dynamic pairs or ranges ahead of time.)

DecisionNetworks.implementation — Function

implementation(dn::DecisionNetwork)

Give the conditional distributions implementing the nodes of a decision network as a NamedTuple mapping node names to distributions.

Instantiating DNs

Every node in DecisionNetworks.jl is a random variable, named with a Symbol. Values and distributions associated with them are passed around as keyword arguments (a pattern seen in many places in Decisions.jl). So, to make a new DN, we pass these distributions into the constructor for a Type{<:DecisionNetwork}. (The Type{<:DecisionNetwork} itself gives the decision graph - which tells us how these random variables condition each other.)

There are several ways to get a Type{<:DecisionNetwork} that can be instantiated. For instance, we might use the premade MDP_DN network type, which underlies Markov decision processes:

my_dn = MDP_DN(;
    sp = (rng; s, a) -> rand(rng),
    r = (rng; s, a, sp) -> s + a
)

MDP_DN (alias for DecisionNetwork{…}):
a  | s         (no dist)
r  | a, s, sp  (AnonymousDist)
sp | a, s      (AnonymousDist)
s  => sp

Tip

Functions and Spaces that are passed to the constructor as implementations for distributions are automatically converted to ConditionalDists.

The pretty-printed output tells us something about the DN: it has nodes named a, r, s, and sp, it has the given conditionings for those nodes, and it's a DDN where sp becomes s at each iterate. Indeed, all Markov decision problems have decision network nodes s (state), sp (successor state), r (reward), and a (action). The distributions for sp, r, and a are the state transition, reward function, and policy respectively. In the parentheses, we see that we've provided implementations for r and sp, but a has no implementation; that is, it is a decision node.

Note that the type MDP_DN - that is, the decision graph - tells us everything there is to know about the network structure of an MDP. The prenamed decision network types defined by DecisionNetworks.jl cover a few commonly used network structures, but later sections in this documentation also deal with defining arbitrary decision graphs.

Sampling DNs

As mentioned above, every node in a DN is a random variable with a Symbol name. Conditional distributions for nodes can be obtained with index notation:

my_dn[:sp]

AnonymousDist: P(::Any | a, s)

This gives the conditional distribution for sp in this DDN (that is, the state transition). All sorts of useful functions for algorithm writers are defined on conditional distributions, documented on their own manual page.

An entire DN or DDN can be sampled given inputs with sample. In particular, the behavior argument is used to provide implementations for all distributions not already implemented in the network: that is, implementations for the action nodes.

DecisionNetworks.sample — Function

sample(fn = (_) -> false, dn::DecisionNetwork [, decisions::NamedTuple, input::NamedTuple, output::Tuple])

Sample nodes or plates out in decision network dn based on input values in and node implementations provided by decisions and dn.implementation. fn, if present, executes upon each iteration of a (dynamic) network on a NamedTuple mapping the names out to their values, and stops when fn returns true.

Returns Terminal() if a terminal condition is reached. Otherwise, returns a NamedTuple mapping the names out to their values after the last (or only) iteration.

Only ancestors of out, up to (but not including) the nodes in in, are sampled. If any of the sampled nodes are not implemented by either dn.implementation or decisions, an error is thrown. If both dn.implementation and decisions specify the same node, the implementation in decisions is preferred.

Decision network traits

Decisions.jl defines a number of traits on various objects, especially decision networks. They are implemented using the Holy trait pattern. The full documentation for all traits is available here.

Traits on decision networks are used to give an exact idea of how particular semantic concepts, like "partial observability" and "sequentiality", correspond to actual decision networks. These traits can be combined to form new networks.

Todo

Traits are not automatically applied to @markov_alias'd networks.

Trait	Possibilities supplied by Decisions.jl	Description

Multiagency	NoAgent, SingleAgent, DefiniteAgent, IndefiniteAgent	Number of agents in a problem (and whether that number is known)

Observability	FullyObservable, PartiallyObservable	Whether input into decision nodes is a sufficient statistic for other nodes in the DN

Centralization	Centralized, Decentralized	Whether input into decision nodes is the same across agents

MemoryPresence	MemoryPresent, MemoryAbsent	Whether there is an agent-defined information aggregator across DDN iterates

RewardStyle	NoReward, SingleReward, DefiniteRewards, IndefiniteRewards	Whether a "reward" node is present, how many rewards it represents, and what it is conditioned on

Statefulness	Stateful, Stateless, AgentFactored	Presence and structure of a "state" node

Sequentiality	Simultaneous, Sequential	Whether decision nodes are calculable in parallel

Cooperation	Cooperative, Competitive, Individual	Whether agents are implied to share or interfere with each others' objectives

AgentCorrelation	Correlated, Uncorrelated	Whether agents jointly or independently sample their decisions

TimestepStyle	FixedTime, SemiMarkov	Whether the network is implied to represent a semi-Markov problem