Advanced: Internals

More structural ConditionalDists

There are some structural ConditionalDists that are primarily used in transformations. These distributions act as wrappers around others, which can cause a significant performance drop as transformations are aggregated. As such, they are considered unstable and are likely to be subject to significant additional refactoring.

DecisionNetworks.RenamedDist — Type

RenamedDist <: ConditionalDist

A conditional distribution that wraps another, renaming the conditioning variables.

DecisionNetworks.MergedDist — Type

MergedDist <: ConditionalDist

A conditional distribution which merges two subdistributions, using one as an input for another.

Tools for traversing DecisionNetworks

These are the tools used at compile time to generate fast sampling code and nonambiguous networks.

DecisionNetworks._standardize_dn_type — Function

_get_dn_type_params(nodes, dynamic_pairs=nothing, ranges=nothing)

Standardize the type parameters for a decision network (imposing an order, making them the correct types, etc.)

Output is a 3-tuple. nodes is required. If dynamic_pairs and ranges are not provided, nothing is returned for them.

DecisionNetworks._crawl_dn — Function

_crawl_dn(dn, constituents, input, output)

Give all nodes required to compute nodes output of dn, given that nodes input are known, in order.

DecisionNetworks._node_order — Function

_node_order(dn, rv)

Calculate the order of the random variable named rv in decision network dn; that is, n such that rv is conditioned on random variables of at most order n-1.

DecisionNetworks._make_node_assignment — Function

_make_node_assignment(dn, id; in_place=false)

Generate an Expr that updates random variable id based on decision network dn.

DecisionNetworks._make_node_initialization — Function

_make_node_initialization(dn, id)

Infer an expression that initializes an empty MArray of the correct size and name for node id in dn.

Random variable groups

DecisionNetworks.RVGroup — Type

abstract type RVGroup{id, idxs}

Abstract base type representing a group of random variables, named id, over axes with indices idxs.

DecisionNetworks.expr — Function

expr(::RVGroup)

Give an indexing expression for a group of random variables that specifies single variables in the group.

Utilities

Simple utilities used across the project:

DecisionNetworks._sortkeys — Function

_sortkeys(::NamedTuple)

Sort the keys in a NamedTuple (to enforce order invariance).

DecisionNetworks._sorted_tuple — Function

_sorted_tuple(v)

Give a tuple containing the elements in v, sorted according to sort.

Useful for imposing a structure on a Tuple used as a type parameter.