Interoperability

Graphs.jl

DecisionNetworks and DecisionGraphs can be converted into the Graphs.jl ecosystem. They are interpreted as SimpleDiGraphs, with the nodes in the same order as given by node_names.

Beware that this strips several important pieces of information, leaving only the underlying graph:

Implementations (conditional distributions) for all nodes

Node names

Independence relationships between plates and their conditions

DecisionNetworks.as_graphs_jl — Function

as_graphs_jl(djl)

Give a Graphs.jl-style directed graph (that is, a SimpleDiGraph) corresponding to the DecisionGraph or DecisionNetwork djl.