Spaces

Spaces represent abstract sets of values represented by a shared Type. They are used to define supports for ConditionalDists and are in general handy to query for many decision-making algorithms.

DecisionNetworks.Space — Type

abstract type Space{T}

Abstract base representation for spaces, possibly infinite sets of instances all backed by type T.

The space interface

Spaces are subtypes of Space{T}. The space interface is related to the iteration interface, but not all spaces can be iterated (some are continuous).

Mandatory functions

Base.in(el, s::Space): Give whether item el is in space s.

Base.eltype(::Space{T}): Gives the backing type of the space. Defaults to T.

Optional functions

Base.zero(::Space): Provide the additive identity element, assuming + is defined over the space.

Base.one(::Space): Provide the multiplicative identity element, assuming * is defined over the space.

Base.length(::Space): Give the cardinality of a space, if it is finite. Otherwise (and by default), return Inf.

Base.iterate(::Space [, state]): For a discrete state, provide iteration support (see (Iteration)[https://docs.julialang.org/en/v1/manual/interfaces/#man-interface-iteration])

Predefined spaces

Some handy Space types are named by Decisions.jl:

DecisionNetworks.FiniteSpace — Type

FiniteSpace{T, N} <: Space{T}

Representation of a finite set backed by type T.

Supports iteration.

DecisionNetworks.RangeSpace — Type

RangeSpace{T} <: Space{T}

Space representing range of set elements backed by type T from lb to ub, inclusive (according to ≤).

DecisionNetworks.TypeSpace — Type

TypeSpace{T} <: Space{T}

Space that is exactly coextensive with its backing type, T: that is, any instance of T is an element of T.

TypeSpace{T} can also be used (with caution) to represent proper subsets of T when the exact extent of the subset is unknown or difficult to calculate.

DecisionNetworks.SingletonSpace — Type

SingletonSpace{T} <: Space{T}

Space that consists of exactly one element el.