Traits

Decisions.jl provides a variety of named traits on DecisionNetworks, ConditionalDists, and other objects in the package. They serve as hints to algorithms as well as the package itself.

This page provides the technical docs for each trait.

DecisionNetworks.DecisionsTrait — Type

abstract type DecisionsTrait

Abstract base trait type for all traits used in Decisions.jl.

Traits on decision networks

Sequentiality

DecisionNetworks.Sequentiality — Type

abstract type Sequentiality <: DecisionsTrait
Sequentiality(::DecisionNetwork)

Abstract trait type denoting whether agents act simultaneously or sequentially.

DecisionNetworks.Sequential — Type

Sequential <: Sequentiality

Trait indicating agents act sequentially (meaning each agent controls one or more independent decision nodes).

DecisionNetworks.Simultaneous — Type

Simultaneous <: Sequentiality

Trait indicating agents act simultaneously (meaning agents implement different indices of the same decision node(s)).

Partial observability / information perfection

DecisionNetworks.Observability — Type

abstract type Observability <: DecisionsTrait
Observability(::DecisionNetwork)

Abstract trait type indicating whether input to decision nodes passed to an agent or agents is "full," in that it is at least as informative as knowing the values of its predecessors.

DecisionNetworks.FullyObservable — Type

FullyObservable <: Observability

Trait indicating decision nodes receive perfect / full information.

DecisionNetworks.PartiallyObservable — Type

PartiallyObservable <: Observability

Trait indicating decision nodes do not receive perfect / full information.

Number of agents

DecisionNetworks.Multiagency — Type

abstract type Multiagency <: DecisionsTrait
Multiagency(::DecisionNetwork)

Abstract trait type denoting the number of agents, if any, in a problem.

DecisionNetworks.NoAgent — Type

struct NoAgents <: Multiagency

Trait denoting a network supports no agents (implying no decision nodes).

DecisionNetworks.SingleAgent — Type

struct SingleAgent <: Multiagency

Trait denoting presence of exactly one agent.

DecisionNetworks.MultiAgent — Type

abstract type MultiAgent <: Multiagency

Abstract trait type denoting a network supports multiple agents.

The exact number of agents may (with DefiniteAgents) or may not (with IndefiniteAgents) be specified.

DecisionNetworks.DefiniteAgents — Type

abstract type DefiniteAgents{N} <: MultiAgent

Trait denoting support for exactly N agents.

DecisionNetworks.IndefiniteAgents — Type

abstract type DefiniteAgents{N} <: MultiAgent

Trait denoting support for some number of agents, which is not known ahead of time.

Agent cooperation

DecisionNetworks.Cooperation — Type

abstract type Cooperation <: DecisionsTrait
Cooperation(::DecisionNetwork)

Abstract trait indiating the extent to which a decision network implies cooperative agents.

DecisionNetworks.Cooperative — Type

Cooperative <: Cooperation

Trait indicating agents are implied to cooperate or receive a benefit from cooperating.

DecisionNetworks.Competitive — Type

Competitive <: Cooperation

Trait indicating agents are implied to compete or receive a benefit from competing.

DecisionNetworks.Individual — Type

Individual <: Cooperation

Trait indicating agents are implied to act independently, neither cooperating nor competing.

Agent memory

DecisionNetworks.MemoryPresence — Type

abstract type MemoryPresence <: DecisionsTrait
MemoryPresence(::DecisionNetwork)

Abstract trait indicating whether iterates of a dynamic decision network are accrued in an agent-defined memory.

DecisionNetworks.MemoryPresent — Type

MemoryPresent <: MemoryPresence

Abstract trait indicating iterates of a dynamic decision network are accrued in an agent-defined memory.

DecisionNetworks.MemoryAbsent — Type

MemoryAbsent <: MemoryPresence

Abstract trait indicating no accrual of information from iterate to iterate of a dynamic decision network in any agent-defined memory.

Reward presence and style

DecisionNetworks.RewardStyle — Type

abstract type RewardStyle{ids} <: DecisionsTrait
RewardStyle(::DecisionNetwork)

Abstract trait indicating whether a decision network has a node analogous to a reward node: a node no others are conditioned on, typically for use in an objective. If so, indicates whether it represents a single or multiple reward, and the other nodes ids it is conditioned on.

DecisionNetworks.NoReward — Type

struct NoReward <: RewardStyle{()}

Trait indicating a decision network has no node analogous to a "reward" node.

DecisionNetworks.SingleReward — Type

SingleReward{ids} <: RewardStyle{ids}

Trait indicating a decision network has a "reward" node producing a single value, which is conditioned on the nodes ids.

DecisionNetworks.MultipleRewards — Type

MultipleRewards{ids} <: RewardStyle{ids}

Abstract trait indicating a decision network has a "reward" node producing multiple values, which is conditioned on the nodes ids.

The number of values produced can be known (DefiniteRewards) or unknown (IndefiniteRewards).

DecisionNetworks.DefiniteRewards — Type

DefiniteRewards{N, ids} <: MultipleRewards{ids}

Trait indicating a decision network has a "reward" node conditioned on ids which produces exactly N values (that is, it is a plate of size N).

DecisionNetworks.IndefiniteRewards — Type

IndefiniteRewards <: MultipleRewards{ids}

Trait indicating a decision network has a "reward" node conditioned on ids which is a plate of unknown size.

Agent centralization

DecisionNetworks.Centralization — Type

Centralization <: DecisionsTrait
Centralization(::DecisionNetwork)

Abstract trait indicating whether agents receive the same input or not.

DecisionNetworks.Centralized — Type

Centralized <: Centralization

Trait indicating agents receive the same input.

DecisionNetworks.Decentralized — Type

Decentralized <: Centralization

Trait indicating agents do not receive the same input.

Statefulness and state structure

DecisionNetworks.Statefulness — Type

abstract type Statefulness <: DecisionsTrait
Statefulness(::DecisionNetwork)

Abstract trait indicating whether a "state" / "state update" node is present, and if so, how it is structured.

DecisionNetworks.Stateful — Type

Stateful <: Statefulness

Trait indicating that a decision network has a "state" / "state update" node.

DecisionNetworks.Stateless — Type

Stateless <: Statefulness

Trait indicating that a decision network has no "state" / "state update" node.

DecisionNetworks.AgentFactored — Type

AgentFactored <: Statefulness

Trait indicating that a decision network has a "state" / "state update" node, and it is a plate with an agent-index axis.

Semi-Markovianness and step style

DecisionNetworks.TimestepStyle — Type

abstract type TimestepStyle <: DecisionsTrait
TimestepStyle(::DecisionNetwork)

Abstract trait indicating the time interpretation of each iterate of a dynamic decision network.

DecisionNetworks.SemiMarkov — Type

SemiMarkov <: TimestepStyle

Trait indicating a dynamic decision network is semi-Markov: the sojourn time from iteration to iteration is given by a node in the network.

DecisionNetworks.FixedTimestep — Type

FixedTimestep <: TimestepStyle

Trait indicating iterates in a dynamic decision network are equal, fixed timesteps with no semi-Markov considerations.

Agent correlation

DecisionNetworks.AgentCorrelation — Type

abstract type AgentCorrelation <: DecisionsTrait
AgentCorrelation(::DecisionNetwork)

Abstract trait indicating whether decision nodes are sampled jointly or independently across agents.

DecisionNetworks.Correlated — Type

Correlated <: AgentCorrelation

Trait indicating all decision nodes are sampled jointly across agents.

DecisionNetworks.Uncorrelated — Type

Uncorrelated <: AgentCorrelation

Trait indicating all decision nodes are sampled independently across agents.

Traits on random variables

RVGroups can accept traits, which are primarily used as hints to sample and other such functions for optimal performance.

Due to their internal role they are not exported by default.

Terminality

DecisionNetworks.Terminality — Type

Terminality <: DecisionsTrait
Terminality(::RVGroup)

Abstract trait on a group of random variables indicating whether the corresponding distribution is expected to be able to produce Terminal() when sampled.

DecisionNetworks.Terminable — Type

Terminable <: Terminality

Trait indicating a group of random variables might produce Terminal() under certain values of its conditioning variables.

DecisionNetworks.NotTerminable — Type

NotTerminable <: Terminality

Trait indicating a group of random variables never produces Terminal() under any circumstances,.

DecisionNetworks.MaybeTerminable — Type

MaybeTerminal <: Terminality

Trait indicating a group of random variables may be able to produce Terminal() under some values of its conditioning variables, or for none of them.

Traits on conditional distributions

Todo

None exist yet. Determinism is the most relevant left to be implemented.

Traits on spaces

Todo

None exist yet. I think things like Finitude and Countability are reasonable, but they're a low priority.