Transformations

One of Decisions.jl's greatest strengths is its ability to flexibly and precisely modify decision networks. Decision problems can be changed from multiple to single agents, full to partial observability, and all sorts of other changes in one line by transforming the underlying decision network.

Such transformations are subtypes of the DNTransformation base class. Transformations can affect only implemented DecisionNetworks, only their DecisionGraphs, or both.

DecisionNetworks.DNTransformation — Type

abstract type DNTransformation

Abstract base class for transformations of decision networks and graphs.

Transformations are callable: (t::DNTransformation)(p) = transform(t, p).

DecisionNetworks.transform — Function

transform(t::DNTransformation, d::DecisionNetwork)
transform(t::DNTransformation, d::DecisionGraph)

Apply transformation t to decision network or graph d, returning a new network or graph.

Provided transformations

DecisionNetworks.Insert — Type

Insert <: DNTransformation
Insert(nodes...)

Insert one or more nodes into a decision graph or decision network.

nodes are supplied as node definitions; that is, pairs of the form (::ConditioningGroup...) => ::Plate.

DecisionNetworks.Implement — Type

Implement <: DNTransformation
Implement(; impls::ConditionalDist...)

Augment a decision network with conditional distribution(s) implementing one or more of its nodes.

Conditional distributions are supplied to nodes matching the names of their kwargs. If a distribution is supplied for a node that has one, the existing distribution is replaced.

DecisionNetworks.Unimplement — Type

Unimplement <: DNTransformation
Unimplement(nodes...)

Remove the conditional distribution(s) implementing one or more nodes in the decision network, if they exist.

The node names are given as Symbols.

DecisionNetworks.Recondition — Type

Recondition <: DNTransformation
Recondition(; nodes...)

Change the conditioning of one or more nodes in a decision graph (not a decision network).

New conditionings are supplied as tuples of ConditioningGroups, with the keyword name giving the node they condition.

DecisionNetworks.IndexExplode — Type

IndexExplode <: DNTransformation
IndexExplode(idx; sep='_')

Split all plates a in a decision network or graph over index idx into nodes a_1, a_2, ..., a_N, where N is given by ranges.

In a decision network, when a node with an implementation dist is exploded along axis i, the resulting subnodes have distributions fix(dist; i=1), fix(dist; i=2), etc.

DecisionNetworks.MergeForward — Type

MergeForward <: DNTransformation
MergeForward(nodes...)

Merges nodes named in nodes forward in a decision network or decision graph: for each such node n, nodes that n as an input now have the inputs of n as inputs (and n is removed from the network).

In a decision network, nodes with implemented distributions have those distributions merged with MergedDist.

DecisionNetworks.Rename — Type

Rename <: DNTransformation
Rename(; names...)

Rename nodes in a decision network.

Each keyword argument maps an old to new node name.

Todo

Many useful transformations are still missing from DecisionNetworks.jl, and those that do exist have yet to be thoroughly tested.

Examples

Transforming a Markov game into an MDP

One might want to interpret a Markov game as an ordinary MDP by assuming some opponent behavior. This is a matter of a few transformations.

We'll start with some Markov game implemented with arbitrary distributions (passed in as Functions and automatically converted). Note that we specify the ranges for the plates of the Markov game - that is, the number of agents - in the type parameter, so this is a two player game.

my_mg = MG_DN{(; i=2)}(;
    sp = (rng; a, s) -> "sp for actions $a",
    r = (rng; a, s, sp, i) -> "r$i"
)

MG_DN (alias for DecisionNetwork{…}):
a[i] | s, i         (no dist)
r[i] | a, s, sp, i  (AnonymousDist)
sp   | a, s         (AnonymousDist)
s    => sp
i    ∈ 1:2

We can turn all the agent-indexed plates into individual nodes with IndexExplode (in this case there are only two apiece for a and r).

my_dn = my_mg |> IndexExplode(:i)

DecisionNetwork{…}:
a_1 | s                (no dist)
a_2 | s                (no dist)
r_1 | a_1, a_2, s, sp  (DecisionNetworks.MergedDist)
r_2 | a_1, a_2, s, sp  (DecisionNetworks.MergedDist)
sp  | a_1, a_2, s      (DecisionNetworks.MergedDist)
s   => sp

This automatically edits the distributions for sp and r to fit the new arrangement. Now, we can provide an implementation for only the opponent policy:

my_dn = my_dn |> Implement(;
    a_2 = (rng; kwargs...) -> "a2"
)

DecisionNetwork{…}:
a_1 | s                (no dist)
a_2 | s                (AnonymousDist)
r_1 | a_1, a_2, s, sp  (DecisionNetworks.MergedDist)
r_2 | a_1, a_2, s, sp  (DecisionNetworks.MergedDist)
sp  | a_1, a_2, s      (DecisionNetworks.MergedDist)
s   => sp

In doing this, we're treating the opponent as part of the model. We can MergeForward the opponent nodes into the state transition (and get rid of the irrelevant opponent reward):

my_dn = my_dn |> MergeForward(:r_2, :a_2)

DecisionNetwork{…}:
a_1 | s           (no dist)
r_1 | a_1, s, sp  (DecisionNetworks.MergedDist)
sp  | a_1, s      (DecisionNetworks.MergedDist)
s   => sp

You may notice that this is not technically a MDP_DN: it has nodes a_1 and r_1, rather than a and r. So, finally, we egoistically rename those nodes:

my_mdp = my_dn|> Rename(; a_1=:a, r_1=:r)

MDP_DN (alias for DecisionNetwork{…}):
a  | s         (no dist)
r  | a, s, sp  (DecisionNetworks.RenamedDist)
sp | a, s      (DecisionNetworks.RenamedDist)
s  => sp

As you can see, this network is indeed the DN for an MDP now.

Transforming a POMDP into an MDP

It is possible for transformations to be lossy: the output network or graph is not necessarily equivalent to the input. This can cause some unexpected behavior. As an example, let's strip the partial observability from a POMDP.

For variety, we'll operate on the DecisionGraph POMDP_DN rather than the DecisionNetwork POMDP_DN(...).

node_names(POMDP_DN)

(:a, :mp, :o, :r, :sp, :m, :s)

Notice that we include the "memory nodes" m and mp by default here. One might expect to be able to simply MergeForward the memory and observation nodes, removing all the partial-observability machinery, to reach a model where a is directly conditioned on s. However, in reality, we see something slightly different:

My_DN = POMDP_DN |> MergeForward(:m, :mp, :o)
conditions(My_DN, :a)

()

Instead of being conditioned on s, a is conditioned on nothing at all! This is because m is an input to the network (except in iterates past the first, where it gets the output of mp), and therefore unconditioned. When it is merged into a, a then has no conditions either.

We can recover the MDP structure by explicitly conditioning the a node:

My_DN = POMDP_DN |> MergeForward(:m, :mp, :o) |> Recondition(; a=(Dense(:s),))
conditions(My_DN, :a)

(:s,)

Finally, let's make a DecisionNetwork with the new DecisionGraph and make sure it's actually an MDP:

My_DN()

MDP_DN (alias for DecisionNetwork{…}):
a  | s         (no dist)
r  | a, s, sp  (no dist)
sp | a, s      (no dist)
s  => sp