Metrics

Decision networks alone have no concepts of solution or objective - they are just environment models. At most, they have action nodes that are not yet implemented, to be determined "somewhere else" then used as input to sample and the like.

DecisionProblems.jl provides the missing link. A decision problem joins a decision network and a metric, which indicates an optimization objective over the random variables in the network. Decision algorithms implement the missing distributions in the decision network in order to optimize that metric.

Using decision metrics

DecisionMetric and its descendants define objectives over decision networks. Decision metrics vary from simple benchmarks ("how many steps did this DDN run?") to ubiquitous objective concepts ("what is the discounted sum of rewards from this state?").

DecisionProblems.DecisionMetric — Type

DecisionMetric

A decision metric: an aggregator of outputs of a DecisionSetting, DecisionEnvironment, or similar object.

When called with one argument, performs aggregate!.

Unlike most Decisions.jl components they are allowed (and expected) to be mutable and stateful.

DecisionMetrics act as aggregators for the values of random variables, and among other things, they can be provided as the first argument of sample. At each iterate (or once after execution, for a non-dynamic decision network), the metric's aggregate! is called with a NamedTuple of random variable values from the network.

DecisionProblems.aggregate! — Function

aggregate!(::DecisionMetric, values)

Aggregate values in the NamedTuple values into the metric.

Unlike most components of Decisions.jl, decision metrics are stateful. (This helps prevent excessive memory usage from computing episode traces). Use output to retrieve the result after aggregation, or reset! to reset the metric object to its initial configuration:

DecisionProblems.output — Function

output(::DecisionMetric)

Give the current value of the aggregated metric calculated by the DecisionMetric.

DecisionProblems.reset! — Function

reset!(::DecisionMetric)

Resets the DecisionMetric to a state before aggregating.

Provided decision metrics

DecisionMetrics are primarily used for consistency and fair comparison: while using a stateful do block for sample is permissible, using a DecisionMetric instead ensures all algorithms which use the same metric follow precisely the same objective definition (even, to some extent, with different decision networks). As such DecisionProblems.jl provides a number of definitive metrics.

Todo

A significant number of metrics remain to be written.

DecisionProblems.Discounted — Type

Discounted{id} <: DecisionMetric

Metric which aggregates a value labelled id as a discounted summation: every subsequent aggregate! increases the discount and decreases the impact of future calls.

Discounting is forward-directed; that is, it is assumed that the first call to aggregate! is the least discounted.

DecisionProblems.DiscountedReward — Type

const DiscountedReward

Alias for Discounted{:r}.

DecisionProblems.Trace — Type

Trace{ids} <: DecisionMetric

Metric which simply aggregates all values labelled in the Tuple ids into Vectors.

If ids is the empty tuple, aggregates all values passed to it.

DecisionProblems.NIters — Type

NIters <: DecisionMetric

A metric that simply tracks the number of times it has been aggregated.

More efficient for computing episode lengths than length(output(t::Trace)).