Decision problems

Decision problems join an objective (that is, a DecisionMetric) with a model (that is, a DecisionNetwork). They also specify initial distributions for random variables in the problem.

Note the distinction between a decision problem (MDP(...)) and the decision network underlying that problem (MDP_DN(...)).

DecisionProblems.DecisionProblem — Type

DecisionProblem

A decision problem, formally stated: a union of a model, given as a DecisionNetwork, an objective, given as a DecisionNetwork, and an optional initial distribution, given as a ConditionalDist with no conditions.

DecisionProblems.objective — Function

objective(dp::DecisionProblem)

Give the DecisionMetric that is the objective of dp.

DecisionProblems.model — Function

model(dp::DecisionProblem)

Give the DecisionNetwork that is the model for dp.

DecisionProblems.initial — Function

initial(dp::DecisionProblem, s::Symbol)
initial(dp::DecisionProblem)

Give an initial value for the random variable s in problem dp (or a NamedTuple for all random variables for which an initialization is specified, if s is not given).

Named decision problems

Like DecisionNetworks, Decisions.jl provides explicit names for some common decision problems:

DecisionProblems.MDP — Type

const MDP = DecisionProblem{<: DiscountedReward, <: MDP_DN}

A Markov decision process.

DecisionProblems.POMDP — Type

const POMDP = DecisionProblem{<: DiscountedReward, <: POMDP_DN}

A partially observable Markov decision process.

DecisionProblems.MG — Type

const MG = DecisionProblem{<: DiscountedReward, <: MG_DN}

A Markov game.

Transforming decision problems

By default, calling a transformation on a DecisionProblem will transform the underlying DecisionNetwork (while leaving the objective unchanged).